Move ObsoleteCommentSyntaxInspectionTests for QuickFixes into relevant files under RD.QuickFixes
